Move Posting type to a separate module
